+++ services/soaf/src/pages/Chat/Chat.tsx
@@ -1,5 +1,5 @@
 import { PageLayout } from "@shared/components";
-import SpeechBubble from "@shared/components/ui/SpeechBubble";
+import SpeechBubble from "@/shared/components/chat/SpeechBubble";
 
 const DUMMY_CHAT = [
   { message: "안녕", sentAt: new Date().toISOString(), userId: 1 },
