+++ services/soaf/src/shared/components/chat/SpeechBubble.tsx
