+++ services/soaf/src/shared/components/chat/index.ts
@@ -1,0 +1,1 @@
+export * from "./SpeechBubble";
