+++ services/soaf/src/shared/components/ui/index.ts
@@ -3,7 +3,6 @@ export * from "./Button.tsx";
 export * from "./Toast.tsx";
 export * from "./Toaster.tsx";
 export * from "./ToggleButton.tsx";
-export * from "./SpeechBubble.tsx";
 export * from "./Badge.tsx";
 export * from "./Input.tsx";
 export * from "./Card.tsx";
